refactor(models): replace custom ops with native PyTorch equivalents
- Replace DopamineVLARMSNorm -> nn.RMSNorm
- Replace PytorchGELUTanh -> nn.GELU(approximate="tanh")
- Remove _expand_mask; pass bool masks directly to SDPA
- Move repeat_kv inside flex_attention branch only;
  SDPA path uses enable_gqa=True instead
- Replace numpy image preprocessing with
  torchvision.transforms.functional
- Use torch.inference_mode() over torch.no_grad()
- Add @torch.compiler.disable() on SigLino._apply
- Fix key_padding_mask inversion in pooling head
- Run ruff format across all files

+++ dopaminevla/models/siglino/siglino/moe.py
@@ -84,7 +84,9 @@ def _run_experts_for_loop(
 
 
 class GroupedExperts(nn.Module):
-    def __init__(self, dim: int, hidden_dim: int, num_experts: int, activation: str = "silu") -> None:
+    def __init__(
+        self, dim: int, hidden_dim: int, num_experts: int, activation: str = "silu"
+    ) -> None:
         super().__init__()
         self.num_experts = num_experts
         self.w1 = nn.Parameter(torch.empty(num_experts, hidden_dim, dim))
@@ -93,7 +95,9 @@ class GroupedExperts(nn.Module):
         self.activation = activation
 
     def forward(self, x: torch.Tensor, num_tokens_per_expert: torch.Tensor) -> torch.Tensor:
-        return _run_experts_for_loop(self.w1, self.w2, self.w3, x, num_tokens_per_expert, self.activation)
+        return _run_experts_for_loop(
+            self.w1, self.w2, self.w3, x, num_tokens_per_expert, self.activation
+        )
 
     def init_weights(self, init_std: float) -> None:
         nn.init.trunc_normal_(self.w1, mean=0.0, std=0.02)
@@ -119,7 +123,9 @@ class TokenChoiceTopKRouter(nn.Module):
         self.route_norm = route_norm
         self.route_scale = route_scale
 
-    def forward(self, x: torch.Tensor, expert_bias: torch.Tensor | None = None) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
+    def forward(
+        self, x: torch.Tensor, expert_bias: torch.Tensor | None = None
+    ) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
         scores = self.gate(x)
         if self.score_func == "sigmoid":
             scores = torch.sigmoid(scores.float())
@@ -190,7 +196,9 @@ class MoE(nn.Module):
         bs, slen, dim = x.shape
         x = x.view(-1, dim)
 
-        top_scores, selected_experts_indices, num_tokens_per_expert = self.router(x, expert_bias=self.expert_bias)
+        top_scores, selected_experts_indices, num_tokens_per_expert = self.router(
+            x, expert_bias=self.expert_bias
+        )
 
         # Reorder tokens by expert
         token_indices_sorted = torch.argsort(selected_experts_indices.view(-1), stable=True)
@@ -210,7 +218,9 @@ class MoE(nn.Module):
         else:
             out = torch.zeros_like(x)
 
-        routed_output = (routed_output.to(torch.float32) * top_scores_sorted.view(-1, 1)).to(x.dtype)
+        routed_output = (routed_output.to(torch.float32) * top_scores_sorted.view(-1, 1)).to(
+            x.dtype
+        )
 
         out = out.scatter_add(dim=0, index=token_indices_expanded, src=routed_output)
         return out.view(bs, slen, dim)
